move store.js to folder redux
