+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import store from './app/store';
+import store from './app/redux/store';
 import './index.css';
 import App from './components/App.js';
